pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "oralis-studio"
version = "1.0.0"
description = "Text-to-speech application powered by VibeVoice."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
dependencies = [
  "torch==2.9.1+rocm6.4",
  "transformers>=4.51.3,<5.0.0",
  "accelerate",
  "llvmlite>=0.40.0",
  "numba>=0.57.0",
  "diffusers",
  "tqdm",
  "numpy",
  "scipy",
  "librosa",
  "uvicorn[standard]",
  "fastapi",
  "requests",
  "python-multipart>=0.0.26",
]

[project.optional-dependencies]
streamingtts = ["transformers==4.51.3"]

[project.scripts]
oralis = "oralis:main"
preprocess-text = "preprocess_text:main"
studio = "web.app:main"

[project.urls]
"Vibevoice-Homepage" = "https://github.com/microsoft/VibeVoice"
"Vibevoice-Bug Tracker" = "https://github.com/microsoft/VibeVoice/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["oralis_studio*", "web*"]

[tool.setuptools]
py-modules = ["oralis", "preprocess_text"]

[[tool.uv.index]]
name = "pytorch-rocm"
url = "https://download.pytorch.org/whl/rocm6.4"

[tool.uv]
environments = ["sys_platform == 'linux'"]

[tool.uv.sources]
torch = [{ index = "pytorch-rocm" }]
torchaudio = [{ index = "pytorch-rocm" }]
torchvision = [{ index = "pytorch-rocm" }]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
